Update and rename remove.js to unwarn.js
